refactor: Move core component to designsystem component
